Add on_attach helper & implemented for lsp-format
Event based helper allows old on_attach function to be split up
